requirements.txt
ffmpeg-python
numpy
pyaudio
pytube